fix: finally fix cpuUsage

+++ include/Modules/CpuUsage.hpp
@@ -15,12 +15,12 @@ namespace Krell {
     namespace Modules {
         class CpuUsage: public IModule {
             public:
-                size_t total;
-                size_t used;
-                size_t free;
+                double total;
+                double used;
+                double free;
 
-                size_t usedPercent = 0;
-                size_t freePercent = 0;
+                double usedPercent = 0;
+                double freePercent = 0;
                 CpuUsage();
                 ~CpuUsage();
                 void refresh() override;
@@ -28,17 +28,17 @@ namespace Krell {
             private:
                 time_t _lastRefresh;
 
-                size_t _user;
-                size_t _nice;
-                size_t _system;
-                size_t _idle;
-                size_t _iowait;
-                size_t _irq;
-                size_t _softirq;
+                double _user;
+                double _nice;
+                double _system;
+                double _idle;
+                double _iowait;
+                double _irq;
+                double _softirq;
 
-                size_t _total;
-                size_t _used;
-                size_t _free;
+                double _total;
+                double _used;
+                double _free;
         };
     }
 }
